Move consolidated treemap test
